+++ src/wavinfo.c
@@ -18,34 +18,19 @@
  */
 
 #include <stdio.h>
-#include <string.h>
-#include <stdint.h>
-#include <inttypes.h>
-#include <libgen.h>
-#include <math.h>
 
 #include "format.h"
 
-static char *
-format_duration(uint64_t duration)
-{
-    uint64_t seconds = duration / 1000;
-    uint64_t fraction = duration % 1000;
-    return g_strdup_printf("%02" PRIu64 ":%02" PRIu64 ":%02" PRIu64 ".%03" PRIu64, seconds / 60 / 60, (seconds / 60) % 60, seconds % 60, fraction);
-}
-
 int cmd_wavinfo(int argc, char *argv[])
 {
-    int i;
-
-    if( argc < 2) {
-        printf( "Usage: %s [file1.wav] [...]\n", basename( argv[0]));
+    if (argc < 2) {
+        printf("Usage: %s [file1.wav] [...]\n", argv[0]);
         return 1;
     }
 
     format_init();
 
-    for( i = 1; i < argc; i++) {
+    for (int i = 1; i < argc; i++) {
         char *error_message = NULL;
         OpenedAudioFile *oaf = format_open_file(argv[i], &error_message);
 
@@ -55,29 +40,7 @@ int cmd_wavinfo(int argc, char *argv[])
             continue;
         }
 
-        char *duration = format_duration((uint64_t)oaf->sample_info.numBytes * 1000 / (uint64_t)oaf->sample_info.avgBytesPerSec);
-
-        printf("File name:      %s\n", argv[i]);
-        printf("File format:    %s\n", oaf->mod->name);
-        if (oaf->details) {
-            printf("Format details: %s\n", oaf->details);
-        }
-        printf("Duration:       %s (%lu samples)\n", duration, oaf->sample_info.numBytes / oaf->sample_info.blockAlign);
-        printf("Format:         %d Hz / %d ch / %d bit",
-                oaf->sample_info.samplesPerSec,
-                oaf->sample_info.channels,
-                oaf->sample_info.bitsPerSample);
-
-        printf("\n");
-
-        //printf("avgBytesPerSec: %d\n", oaf->sample_info.avgBytesPerSec);
-        //printf("blockAlign:     %d\n", oaf->sample_info.blockAlign);
-        //printf("numBytes:       %lu\n", oaf->sample_info.numBytes);
-        //printf("blockSize:      %d\n", oaf->sample_info.blockSize);
-
-        // TODO: Maybe dump extra metadata fields
-
-        g_free(duration);
+        format_print_file_info(oaf);
         format_close_file(oaf);
 
         printf("\n");